// src/api/generated/models/LeadResponse.ts
export interface LeadResponse {
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    leadId?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    leadName?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    companyName?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    contactName?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    phone?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    leadSource?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    leadStatus?: string;
    /**
     * 
     * @type {number}
     * @memberof LeadResponse
     */
    leadScore?: number;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    industry?: string;
    /**
     * 
     * @type {number}
     * @memberof LeadResponse
     */
    estimatedRevenue?: number;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    assignedUserId?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    assignedUserName?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    convertedCustomerId?: string;
    /**
     * 
     * @type {Date}
     * @memberof LeadResponse
     */
    convertedDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof LeadResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof LeadResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof LeadResponse
     */
    updatedDate?: Date;
}